refactor: centralize Newton-Armijo optimisation
